Move methods to superclass and rename as necessary
